refactor: consolidate types into single types.ts + add public type exports
File structure before:
  src/
    index.ts
    constants.ts
    interfaces/
      currency.ts          (19 LOC, 2 interfaces)
      NumberProperties.ts  (5 LOC, 1 interface)

File structure after:
  src/
    index.ts
    constants.ts
    types.ts               (one file, all 3 interfaces, JSDoc'd)

Why:
  - 3 files instead of 4 -- less to grep through.
  - Conventional `types.ts` name.
  - Same-level imports (`./types`) instead of `./interfaces/...`.
  - PascalCase filename `NumberProperties.ts` was inconsistent with
    everything-else lowercase.

Also adds a public type re-export from index.ts so consumers can:

  import { Tafgeet, Currency, Currencies, NumberProperties } from 'tafgeet-arabic';

This is a NEW public API (additive, no breakage). Previously
consumers had no way to import the types at all -- they'd have to
deep-import from `tafgeet-arabic/dist/src/interfaces/currency`,
which is an anti-pattern and not part of the supported API.

Verified:
  - 107 mocha tests pass.
  - JS compat smoke test passes (validates the published dist/ from
    a plain CommonJS consumer).
  - The interfaces/ directory had no public consumers (it was never
    re-exported from index.ts), so removing it is safe.
