--- src/interfaces/NumberProperties.ts
@@ -1,5 +1,0 @@
-export interface NumberProperties {
-  singular: string;
-  binary: string;
-  plural: string;
-}
