--- src/interfaces/currency.ts
@@ -1,19 +1,0 @@
-export interface Currencies {
-  SDG: Currency;
-  SAR: Currency;
-  QAR: Currency;
-  AED: Currency;
-  EGP: Currency;
-  KWD: Currency;
-  USD: Currency;
-  AUD: Currency;
-  TND: Currency;
-  TRY: Currency;
-}
-export interface Currency {
-  singular: string;
-  plural: string;
-  fraction: string;
-  fractions: string;
-  decimals: number;
-}
